fix(theoros): Share code for building the calldata
