+++ rust/theoros/src/handlers/mod.rs
@@ -1,2 +1,3 @@
+pub mod build_calldata;
 pub mod rest;
 pub mod websocket;
